# niyam_retail/docker/point_of_sale/openapi.yaml
openapi: 3.0.3
info:
  title: Point of Sale API
  description: |
    POS API for retail checkout operations. Supports sessions, transactions, 
    returns, gift cards, cash management, and offline sync.
  version: 1.0.0
  contact:
    name: Vruksha Support
servers:
  - url: http://localhost:8800
    description: Development server
...
paths:
...
  /transactions:
    get:
      tags: [Transactions]
      summary: List transactions
      parameters:
        - name: session_id
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: List of transactions
    
    post:
      tags: [Transactions]
      summary: Create a transaction (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  transaction:
                    $ref: '#/components/schemas/Transaction'
                  receipt_number:
                    type: string
...
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: string
        session_id:
          type: string
        receipt_number:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/TransactionItem'
        subtotal:
          type: number
        tax_total:
          type: number
        discount_total:
          type: number
        total:
          type: number
        payments:
          type: array
          items:
            $ref: '#/components/schemas/Payment'
        customer_id:
          type: string
        created_at:
          type: string
          format: date-time
    TransactionItem:
      type: object
      properties:
        product_id:
          type: string
        sku:
          type: string
        name:
          type: string
        quantity:
          type: integer
        price:
          type: number
        tax_rate:
          type: number
        discount:
          type: number
        line_total:
          type: number
    Payment:
      type: object
      properties:
        method:
          type: string
          enum: [cash, card, upi, gift_card, store_credit]
        amount:
          type: number
        reference:
          type: string
...
    CreateTransactionRequest:
      type: object
      required: [session_id, items, payments]
      properties:
        session_id:
          type: string
        items:
          type: array
          items:
            type: object
            required: [product_id, quantity, price]
            properties:
              product_id:
                type: string
              sku:
                type: string
              name:
                type: string
              quantity:
                type: integer
                minimum: 1
              price:
                type: number
              tax_rate:
                type: number
              discount:
                type: number
        payments:
          type: array
          items:
            type: object
            required: [method, amount]
            properties:
              method:
                type: string
              amount:
                type: number
              reference:
                type: string
        customer_id:
          type: string
        notes:
          type: string